requirements.txt
pandas~=2.3.3
pillow~=11.3.0
PyPDF2~=3.0.1
pytesseract~=0.3.13
python-docx~=1.2.0
scikit-learn~=1.6.1
torch~=2.8.0